Some appearence changes

+++ app/src/main/java/io/svechnikov/telegramchart/chart/data/NavigationBounds.java
@@ -30,8 +30,8 @@ public class NavigationBounds implements Parcelable {
 
     @Override
     public void writeToParcel(Parcel dest, int flags) {
-        dest.writeDouble(left);
-        dest.writeDouble(right);
+        dest.writeFloat(left);
+        dest.writeFloat(right);
     }
 
     @Override
